moved sockets
changed where the socket is instantiated, its now instantiated at index.js since it is the first js file to be service when the page is served. Added new socket.on methods to accomadate users trying to rejoin a game when they refresh the page. However everything needs to be rerendered. when re freshing a page, since react does not remember anything.So we need to save the state of the game on the client in some form, most likely using sessionStorage