override data on a client's session
fixed an issue where a clients session data was not being overwritten if they received a socket.on("deadRoom"), still needs work. If the user happends to be in waitingRoom.js during this and receives this message , errors are thrown as the values are null. To overcome this we can just redirect a user to the home page.